--- src/tailwind.config.js
@@ -1,9 +1,0 @@
-module.exports = {
-  content: [
-    "*"
-  ],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-}
