+++ tailwind.config.js
@@ -1,9 +1,9 @@
-/** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: ['*.{js,jsx}'],
+  content: [
+    "./src/**/*.{js,jsx,ts,tsx}"
+  ],
   theme: {
     extend: {},
   },
   plugins: [],
 }
-
